+++ frejaorgid.js
@@ -10,7 +10,6 @@ const defaultSettings = {
         client_cert: '',
         ca_cert: fs.readFileSync(`./certs/bankid_prod.ca`),
         jwt_cert: fs.readFileSync(`./certs/frejaeid_prod.jwt`),
-        minimumLevel: 'EXTENDED',
         password: '',
         default_country: 'SE'
     },
@@ -19,7 +18,6 @@ const defaultSettings = {
         client_cert: fs.readFileSync('./certs/frejaeid_test.pfx'),
         ca_cert: fs.readFileSync(`./certs/frejaeid_test.ca`),
         jwt_cert: fs.readFileSync(`./certs/frejaeid_test.jwt`),
-        minimumLevel: 'EXTENDED',
         password: 'test',
         default_country: 'SE'
     }        
@@ -44,45 +42,28 @@ function initialize(settings) {
     });
 }
 
-function unPack(self,data) {
-    if (typeof data === 'string') {
-        return { ssn: data, country: self.settings.default_country };
-    } else {
-        return { ssn: data.ssn ? data.ssn : data.toString(), country: data.country ? data.country : self.settings.default_country};
-    }
-}
-
-
 // Lets structure a call for a auth request and return the worker
-async function initAuthRequest(ssn) {
-    var ssn = unPack(this,ssn);
-    return await initRequest(this,'authentication/1.0/initAuthentication', "initAuthRequest="+Buffer.from(JSON.stringify({
-        attributesToReturn: ["EMAIL_ADDRESS", "RELYING_PARTY_USER_ID", "BASIC_USER_INFO"],
+async function initAuthRequest(orgid) {
+    return await initRequest(this,'organisation/authentication/1.0/init', "initAuthRequest="+Buffer.from(JSON.stringify({
+        attributesToReturn: ["EMAIL_ADDRESS", "RELYING_PARTY_USER_ID", "BASIC_USER_INFO", "SSN"],
         minRegistrationLevel: this.settings.minimumLevel,
-        userInfoType: "SSN",
-        userInfo: Buffer.from(JSON.stringify({
-            country: ssn.country,
-            ssn: ssn.ssn
+        userInfoType: "ORG_ID",
+        userInfo: orgid
         })).toString('base64')
-    })).toString('base64')
     );
 }
 
 // Lets structure a call for a sign request and return the worker
-async function initSignRequest(ssn,text) {
-    var ssn = unPack(this,ssn);
-    return await initRequest(this,'sign/1.0/initSignature', "initSignRequest="+Buffer.from(JSON.stringify({
-        attributesToReturn: ["EMAIL_ADDRESS", "RELYING_PARTY_USER_ID", "BASIC_USER_INFO"],
+async function initSignRequest(orgid,text) {
+    return await initRequest(this,'organisation/sign/1.0/init', "initSignRequest="+Buffer.from(JSON.stringify({
+        attributesToReturn: ["EMAIL_ADDRESS", "RELYING_PARTY_USER_ID", "BASIC_USER_INFO", "SSN"],
         minRegistrationLevel: this.settings.minimumLevel,
-        userInfoType: "SSN",
+        userInfoType: "ORG_ID",
         signatureType: 'SIMPLE',
         dataToSignType: 'SIMPLE_UTF8_TEXT',
         dataToSign: { text: Buffer.from(text).toString('base64') },
-        userInfo: Buffer.from(JSON.stringify({
-            country: ssn.country,
-            ssn: ssn.ssn
+        userInfo: orgid
         })).toString('base64')
-    })).toString('base64')
     );
 }
 
@@ -125,14 +106,14 @@ async function initRequest(self,endpoint,data) {
 
 // Check the status of an existing auth request
 async function pollAuthStatus(id,self=this) {
-    return pollStatus(self,'authentication/1.0/getOneResult',"getOneAuthResultRequest="+Buffer.from(JSON.stringify({
+    return pollStatus(self,'organisation/authentication/1.0/getOneResult',"getOneAuthResultRequest="+Buffer.from(JSON.stringify({
         authRef: id
     })).toString('base64'))
 }
 
 // Check the status of an existing sign request
 async function pollSignStatus(id,self=this) {
-    return pollStatus(self,'sign/1.0/getOneResult',"getOneSignResultRequest="+Buffer.from(JSON.stringify({
+    return pollStatus(self,'organisation/sign/1.0/getOneResult',"getOneSignResultRequest="+Buffer.from(JSON.stringify({
         signRef: id
     })).toString('base64'))
 }
@@ -183,12 +164,14 @@ async function pollStatus(self,endpoint,data) {
                 return {
                     status: 'completed', 
                     user: {
-                        ssn: userInfo.ssn,
+                        ssn: userInfo.requestedAttributes.ssn.ssn,
                         firstname: result.data.requestedAttributes.basicUserInfo.name,
                         surname: result.data.requestedAttributes.basicUserInfo.surname,
                         fullname: result.data.requestedAttributes.basicUserInfo.name + ' ' + result.data.requestedAttributes.basicUserInfo.surname
                     },
                     extra: {
+                        country: userInfo.requestedAttributes.ssn.country,
+                        org_id: userInfo.requestedAttributes.organisationIdIdentifier,
                         jwt_token: result.data.details
                     }};                
             default:
@@ -243,7 +226,7 @@ async function followRequest(self, pollMethod, initresp, initcallback=undefined,
 
 // Lets cancel pending requests, dont really care about the results here
 async function cancelAuthRequest(id) {
-      await to(this.axios.post(`${this.settings.endpoint}/authentication/1.0/cancel`, 
+      await to(this.axios.post(`${this.settings.endpoint}/organisation/authentication/1.0/cancel`, 
         "cancelAuthRequest="+Buffer.from(JSON.stringify({
             authRef: id
         })).toString('base64')
@@ -253,7 +236,7 @@ async function cancelAuthRequest(id) {
 
 // Lets cancel pending requests, dont really care about the results here
 async function cancelSignRequest(id) {
-    await to(this.axios.post(`${this.settings.endpoint}/sign/1.0/cancel`, 
+    await to(this.axios.post(`${this.settings.endpoint}/organisation/sign/1.0/cancel`, 
       "cancelSignRequest="+Buffer.from(JSON.stringify({
           signRef: id
       })).toString('base64')